fix(app-container): 修复 Header 格式，合并 Container 和 Application
